# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 班级网站网站 API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /quizzes/start:
    post:
      summary: 随机抽题或按题目 ID 创建练习快照
      description: 随机抽题默认排除往年真题；传 questionIds 时按给定顺序自由组卷。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizStartRequest" }
      responses:
        "201":
          description: 已生成不含答案的答题快照
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizStartResponse" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /quizzes/attempts/active:
    get:
      summary: 获取当前成员最多五个未过期答题草稿或评分任务
      responses:
        "200":
          description: 按最近保存时间倒序返回的可恢复答题
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizActiveAttemptList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /quizzes/attempts/{attemptId}:
    get:
      summary: 仅由所有者读取答题草稿、评分状态或既有结果
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 当前答题生命周期快照
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizAttemptLifecycle" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /quizzes/attempts/{attemptId}/draft:
    patch:
      summary: 按 revision 保存答案和位置并将草稿有效期续至七天后
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizDraftSaveRequest" }
      responses:
        "200":
          description: 草稿已保存并递增 revision
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizDraftSaveResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /quizzes/attempts/{attemptId}/abandon:
    post:
      summary: 幂等放弃未提交答题；评分中或已提交答题拒绝放弃
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
      responses:
        "201":
          description: 答题已放弃或此前已放弃
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizAbandonResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /quizzes/questions:
    get:
      summary: 分页浏览整个题库
      description: 返回题干、选项和配图，但不返回答案、解析或简答题评分标准。
      parameters:
        - in: query
          name: subjectId
          schema: { type: string }
        - in: query
          name: chapterIds
          style: form
          explode: false
          schema:
            type: array
            minItems: 1
            maxItems: 20
            items: { type: string }
        - in: query
          name: chapterMatch
          schema: { $ref: "#/components/schemas/ChapterMatch" }
        - in: query
          name: includeCrossChapter
          schema: { type: boolean }
        - in: query
          name: typeLabel
          schema: { type: string, maxLength: 100 }
        - in: query
          name: source
          schema: { $ref: "#/components/schemas/QuizQuestionSourceFilter" }
        - in: query
          name: pastPaper
          schema: { type: string, enum: [ALL, EXCLUDE, ONLY], default: ALL }
        - in: query
          name: paperId
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string, maxLength: 200 }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: 无答案题目分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizLibraryResponse" }
    post:
      summary: 创建一道题目
      x-required-roles: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizQuestionWriteRequest" }
      responses:
        "201": { description: 已创建 }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /quizzes/questions/{id}:
    get:
      summary: 获取包含答案与评分标准的题目编辑数据
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 题目编辑数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizQuestionEditorData" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: 更新题目及其多章节关联
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizQuestionWriteRequest" }
      responses:
        "200":
          description: 题目已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizQuestionEditorData" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      summary: 停用题目并保留历史答题快照
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 题目已停用；重复调用保持幂等成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizQuestionDeleteResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes/questions/{id}/images:
    post:
      summary: 上传并关联一道题目的配图
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                caption: { type: string, maxLength: 300 }
                sortOrder: { type: integer, minimum: 0, maximum: 10000 }
      responses:
        "201":
          description: 配图已关联
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizImage" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes/questions/{questionId}/images/{photoId}:
    delete:
      summary: 解除题目与配图的关联
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
        - in: path
          name: photoId
          required: true
          schema: { type: string }
      responses:
        "200": { description: 已解除关联 }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes/papers:
    get:
      summary: 获取往年真题试卷列表
      parameters:
        - in: query
          name: subjectId
          schema: { type: string }
      responses:
        "200":
          description: 往年真题列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizPaperListResponse" }
  /quizzes/history:
    get:
      summary: 获取当前成员已提交的答题历史
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 按创建时间倒序返回的答题历史
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizHistoryResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /quizzes/wrong:
    get:
      summary: 查询当前成员的错题索引
      description: 每道错题保留累计答错次数、最近错误时间、最近题目快照及简答题最近错误反馈；答对不会删除历史错题。
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: subjectId
          schema: { type: string }
        - in: query
          name: chapterIds
          description: 逗号分隔的章节 ID，最多 20 个
          schema: { type: string }
        - in: query
          name: chapterMatch
          schema: { $ref: "#/components/schemas/ChapterMatch" }
        - in: query
          name: includeCrossChapter
          schema: { type: boolean, default: false }
        - in: query
          name: typeLabel
          schema: { type: string, maxLength: 100 }
        - in: query
          name: source
          schema: { $ref: "#/components/schemas/QuizQuestionSourceFilter" }
        - in: query
          name: search
          schema: { type: string, maxLength: 200 }
      responses:
        "200":
          description: 最近错题列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizWrongResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /quizzes/papers/{paperId}/start:
    post:
      summary: 按试卷顺序开始整套往年真题测试
      parameters:
        - in: path
          name: paperId
          required: true
          schema: { type: string }
      responses:
        "201":
          description: 已生成整套试卷答题快照
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizStartResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes/{attemptId}/submit:
    post:
      summary: 按不可变快照和规范化答案幂等提交并自动判分
      parameters:
        - in: path
          name: attemptId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizSubmitRequest" }
      responses:
        "200":
          description: 已完成评分；相同 submissionHash 重复提交返回同一结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizSubmitCompletedResponse" }
        "202":
          description: 相同答案正在由持有租约的请求评分
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizSubmitPendingResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bmc3_session
  parameters:
...
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
...
  responses:
...
    Unauthorized:
      description: 未登录、会话失效或凭据无效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: 当前账号无权限或操作被安全规则拒绝
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: 目标不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: 操作与当前资源状态冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: 请求体校验失败
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    ChapterMatch:
      type: string
      enum: [ANY, ALL]
      default: ANY
...
    QuizQuestionSourceFilter:
      type: string
      enum: [AI, NON_AI]
...
    ApiError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        code: { type: string }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        requestId: { type: string }
...
    QuestionType:
      type: string
      enum: [SINGLE, MULTIPLE, TRUE_FALSE, SHORT_ANSWER]
...
    QuizImage:
      type: object
      required: [id, caption, mimeType, size, width, height, url]
      properties:
        id: { type: string }
        caption: { type: string }
        mimeType: { type: string }
        size: { type: integer, minimum: 0 }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
        url: { type: string }
...
    QuizStartRequest:
      type: object
      additionalProperties: false
      properties:
        count: { type: integer, minimum: 1, maximum: 100 }
        subjectId: { type: string }
        chapterIds:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items: { type: string }
        chapterMatch: { $ref: "#/components/schemas/ChapterMatch" }
        includeCrossChapter: { type: boolean, default: false }
        typeLabels:
          type: array
          minItems: 1
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 100 }
        types:
          description: 兼容旧客户端的判分类型过滤；新客户端使用 typeLabels。
          type: array
          minItems: 1
          maxItems: 4
          items: { $ref: "#/components/schemas/QuestionType" }
        source: { $ref: "#/components/schemas/QuizQuestionSourceFilter" }
        includePastPapers: { type: boolean, default: false }
        questionIds:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items: { type: string }
      anyOf:
        - required: [count]
        - required: [questionIds]
    QuizStartResponse:
      type: object
      required: [attemptId, questions]
      properties:
        attemptId: { type: string }
        questions:
          type: array
          items: { $ref: "#/components/schemas/QuizQuestionSummary" }
    QuizAttemptStatus:
      type: string
      enum: [DRAFT, SCORING, SUBMITTED, SCORING_FAILED, ABANDONED]
    QuizAnswerMap:
      type: object
      additionalProperties:
        type: array
        items: { type: string }
    QuizAttemptLifecycle:
      type: object
      additionalProperties: false
      required: [attemptId, status, revision, position, answers, questions, savedAt, expiresAt, abandonedAt, submittedAt, score, total, results, gradingError]
      properties:
        attemptId: { type: string }
        status: { $ref: "#/components/schemas/QuizAttemptStatus" }
        revision: { type: integer, minimum: 0 }
        position: { type: integer, minimum: 0 }
        answers: { $ref: "#/components/schemas/QuizAnswerMap" }
        questions:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/QuizQuestionSummary" }
        savedAt: { type: [string, "null"], format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }
        abandonedAt: { type: [string, "null"], format: date-time }
        submittedAt: { type: [string, "null"], format: date-time }
        score: { type: [number, "null"], minimum: 0 }
        total: { type: integer, minimum: 0 }
        results:
          oneOf:
            - type: array
              items: { type: object, additionalProperties: true }
            - { type: "null" }
        gradingError: { type: [string, "null"], maxLength: 500 }
    QuizActiveAttemptList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 5
          items: { $ref: "#/components/schemas/QuizAttemptLifecycle" }
    QuizDraftSaveRequest:
      type: object
      additionalProperties: false
      required: [revision, position, answers]
      properties:
        revision: { type: integer, minimum: 0 }
        position: { type: integer, minimum: 0 }
        answers: { $ref: "#/components/schemas/QuizAnswerMap" }
    QuizDraftSaveResponse:
      type: object
      additionalProperties: false
      required: [attemptId, revision, position, savedAt, expiresAt]
      properties:
        attemptId: { type: string }
        revision: { type: integer, minimum: 1 }
        position: { type: integer, minimum: 0 }
        savedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    QuizAbandonResponse:
      type: object
      additionalProperties: false
      required: [attemptId, abandoned]
      properties:
        attemptId: { type: string }
        abandoned: { type: boolean, const: true }
        abandonedAt: { type: string, format: date-time }
    QuizSubmitRequest:
      type: object
      additionalProperties: false
      required: [answers]
      properties:
        answers: { $ref: "#/components/schemas/QuizAnswerMap" }
    QuizSubmitPendingResponse:
      type: object
      additionalProperties: false
      required: [attemptId, status, pending]
      properties:
        attemptId: { type: string }
        status: { type: string, const: SCORING }
        pending: { type: boolean, const: true }
    QuizSubmitCompletedResponse:
      type: object
      additionalProperties: false
      required: [attemptId, status, pending, score, total, results]
      properties:
        attemptId: { type: string }
        status: { type: string, const: SUBMITTED }
        pending: { type: boolean, const: false }
        score: { type: number, minimum: 0 }
        total: { type: integer, minimum: 0 }
        results:
          type: array
          items: { type: object, additionalProperties: true }
        submittedAt: { type: [string, "null"], format: date-time }
...
    QuizLibraryResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/QuizQuestionSummary" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
    QuizPaperListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/QuizPaperSummary" }
        total: { type: integer, minimum: 0 }
...
    QuizHistoryResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/QuizHistoryItem" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
...
    QuizWrongResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/QuizWrongItem" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
...
    QuizQuestionWriteRequest:
      type: object
      additionalProperties: false
      required:
        [
          gradingType,
          subjectId,
          chapterIds,
          prompt,
          options,
          correctAnswer,
          explanation,
        ]
      properties:
        gradingType: { $ref: "#/components/schemas/QuestionType" }
        typeLabel: { type: string, minLength: 1, maxLength: 100 }
        subjectId: { type: string }
        chapterIds:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items: { type: string }
        category: { $ref: "#/components/schemas/QuizQuestionCategory" }
        prompt: { type: string, minLength: 2, maxLength: 10000 }
        options:
          type: array
          items: { $ref: "#/components/schemas/QuizOption" }
        correctAnswer: { type: array, items: { type: string } }
        gradingRubric: { $ref: "#/components/schemas/GradingRubric" }
        explanation: { type: string, maxLength: 10000 }
        pastPaperId: { type: string }
        paperOrder: { type: integer, minimum: 1, maximum: 10000 }
    QuizQuestionEditorData:
      allOf:
        - { $ref: "#/components/schemas/QuizQuestionSummary" }
        - type: object
          required: [correctAnswer, explanation]
          properties:
            correctAnswer: { type: array, items: { type: string } }
            gradingRubric: { $ref: "#/components/schemas/GradingRubric" }
            explanation: { type: string }
    QuizQuestionDeleteResult:
      type: object
      required: [id, deleted]
      properties:
        id: { type: string }
        deleted: { type: boolean, const: true }
...